Arduino_Radar/pyproject.toml
[tool.poetry]
name = "radar"
version = "0.0.1"
description = "Interface para o sensor ultrassonico Arduino"
authors = ["Gabriel Coelho"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
matplotlib = "^3.7.1"
pyserial = "^3.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"